requirements.txt
astroid==1.6.1
colorama==0.3.9
Django==2.0.2
django-cors-headers==2.1.0
djangorestframework==3.7.7
isort==4.3.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.1
pylint==1.8.2
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
wrapt==1.10.11
